List of names of administrative areas form the GADM data base

A dataset containing names of all the available administrative areas from the GADM data base. Names describe distinct administrative areas in five levels.

adm_area_names

Format

A data frame with 41535 rows and 4 columns.

ISO3

character, country code.

NAME_0

character, name of administrative areas at level 0.

NAME_1

character, name of administrative areas at level 1.

NAME_2

character, name of administrative areas at level 2.

Source

https://gadm.org/

Examples

data("adm_area_names", package = "rangemap")
adm_area_names[1:10, 1:4]

#>    ISO3      NAME_0     NAME_1          NAME_2
#> 1   AFG Afghanistan Badakhshan         Baharak
#> 2   AFG Afghanistan Badakhshan          Darwaz
#> 3   AFG Afghanistan Badakhshan        Fayzabad
#> 4   AFG Afghanistan Badakhshan       Ishkashim
#> 5   AFG Afghanistan Badakhshan            Jurm
#> 6   AFG Afghanistan Badakhshan         Khwahan
#> 7   AFG Afghanistan Badakhshan          Kishim
#> 8   AFG Afghanistan Badakhshan Kuran Wa Munjan
#> 9   AFG Afghanistan Badakhshan            Ragh
#> 10  AFG Afghanistan Badakhshan   Shahri Buzurg

# Country names: Level 0
unique(adm_area_names$NAME_0)

#>   [1] Afghanistan                      Akrotiri and Dhekelia           
#>   [3] Albania                          Algeria                         
#>   [5] American Samoa                   Angola                          
#>   [7] Argentina                        Australia                       
#>   [9] Austria                          Azerbaijan                      
#>  [11] Bangladesh                       Belarus                         
#>  [13] Belgium                          Benin                           
#>  [15] Bhutan                           Bolivia                         
#>  [17] Bosnia and Herzegovina           Botswana                        
#>  [19] Brazil                           Brunei                          
#>  [21] Bulgaria                         Burkina Faso                    
#>  [23] Burundi                          Cambodia                        
#>  [25] Cameroon                         Canada                          
#>  [27] Central African Republic         Chad                            
#>  [29] Chile                            China                           
#>  [31] Colombia                         Costa Rica                      
#>  [33] Côte d'Ivoire                    Croatia                         
#>  [35] Cuba                             Czech Republic                  
#>  [37] Democratic Republic of the Congo Denmark                         
#>  [39] Djibouti                         Dominican Republic              
#>  [41] Ecuador                          Egypt                           
#>  [43] El Salvador                      Equatorial Guinea               
#>  [45] Eritrea                          Estonia                         
#>  [47] Ethiopia                         Faroe Islands                   
#>  [49] Fiji                             Finland                         
#>  [51] France                           French Guiana                   
#>  [53] Gabon                            Gambia                          
#>  [55] Georgia                          Germany                         
#>  [57] Ghana                            Greece                          
#>  [59] Guadeloupe                       Guatemala                       
#>  [61] Guinea-Bissau                    Guinea                          
#>  [63] Guyana                           Haiti                           
#>  [65] Honduras                         Hungary                         
#>  [67] Iceland                          India                           
#>  [69] Indonesia                        Iran                            
#>  [71] Iraq                             Isle of Man                     
#>  [73] Italy                            Japan                           
#>  [75] Jordan                           Kazakhstan                      
#>  [77] Kenya                            Kosovo                          
#>  [79] Kyrgyzstan                       Laos                            
#>  [81] Latvia                           Lebanon                         
#>  [83] Liberia                          Lithuania                       
#>  [85] Luxembourg                       Macao                           
#>  [87] Madagascar                       Malawi                          
#>  [89] Malaysia                         Mali                            
#>  [91] Malta                            Martinique                      
#>  [93] Mauritania                       Mexico                          
#>  [95] Mongolia                         Morocco                         
#>  [97] Mozambique                       Myanmar                         
#>  [99] Namibia                          Nepal                           
#> [101] Netherlands                      New Caledonia                   
#> [103] New Zealand                      Nicaragua                       
#> [105] Niger                            Nigeria                         
#> [107] North Korea                      Norway                          
#> [109] Oman                             Pakistan                        
#> [111] Palestina                        Panama                          
#> [113] Papua New Guinea                 Paraguay                        
#> [115] Peru                             Philippines                     
#> [117] Poland                           Portugal                        
#> [119] Republic of Congo                Reunion                         
#> [121] Romania                          Russia                          
#> [123] Rwanda                           Saint Helena                    
#> [125] Samoa                            São Tomé and Príncipe           
#> [127] Senegal                          Serbia                          
#> [129] Sierra Leone                     Slovakia                        
#> [131] Slovenia                         Solomon Islands                 
#> [133] Somalia                          South Africa                    
#> [135] South Korea                      South Sudan                     
#> [137] Spain                            Sri Lanka                       
#> [139] Sudan                            Suriname                        
#> [141] Swaziland                        Sweden                          
#> [143] Switzerland                      Syria                           
#> [145] Taiwan                           Tajikistan                      
#> [147] Tanzania                         Thailand                        
#> [149] Timor-Leste                      Togo                            
#> [151] Trinidad and Tobago              Tunisia                         
#> [153] Turkey                           Uganda                          
#> [155] Ukraine                          United Arab Emirates            
#> [157] United Kingdom                   United States                   
#> [159] Uruguay                          Uzbekistan                      
#> [161] Vanuatu                          Venezuela                       
#> [163] Vietnam                          Virgin Islands, U.S.            
#> [165] Wallis and Futuna                Yemen                           
#> [167] Zambia                           Zimbabwe                        
#> 256 Levels: Afghanistan Akrotiri and Dhekelia Åland Albania ... Zimbabwe

# Provinces of a country: Level 1
as.character(unique(adm_area_names[adm_area_names$NAME_0 == "Ecuador", "NAME_1"]))

#>  [1] "Azuay"                          "Bolivar"                       
#>  [3] "Cañar"                          "Carchi"                        
#>  [5] "Chimborazo"                     "Cotopaxi"                      
#>  [7] "El Oro"                         "Esmeraldas"                    
#>  [9] "Galápagos"                      "Guayas"                        
#> [11] "Imbabura"                       "Loja"                          
#> [13] "Los Rios"                       "Manabi"                        
#> [15] "Morona Santiago"                "Napo"                          
#> [17] "Orellana"                       "Pastaza"                       
#> [19] "Pichincha"                      "Santa Elena"                   
#> [21] "Santo Domingo de los Tsachilas" "Sucumbios"                     
#> [23] "Tungurahua"                     "Zamora Chinchipe"              

# Some administrative areas at level 2 of India
l2_India <- as.character(unique(adm_area_names[adm_area_names$NAME_0 == "India",
                                               "NAME_2"]))