rangemap: Simple Tools for Defining Species Ranges

rangemap contains a collection of tools to create species range maps based on occurrence data, statistics, and spatial objects. Other tools in this collection can be used to analyze the environmental characteristics of the species ranges. Plotting options to represent results in various manners are also available. Results obtained using these tools can be used to explore the distribution of species and define areas of occupancy and extent of occurrence of species.

rangemap main functions

rangemap_boundaries, rangemap_buffer, rangemap_enm, rangemap_explore, rangemap_plot, rangemap_hull, rangemap_tsa, ranges_emaps, ranges_espace